UserServer 포트 검증과 종료 경로 통일 및 매칭 중단 이름 정리

+++ network/Core/Listener.cs
@@ -20,7 +20,7 @@ public sealed class Listener(ILogger<Listener> logger)
     private Task? _listenTask;
     public event NewClientHandler? ClientConnected;
 
-    public void Start(IPAddress address, short port)
+    public void Start(IPAddress address, int port)
     {
         lock (_lifecycleLock)
         {
